feat: migrate CommonJS module to ES module syntax
